+++ UCR.Core/Models/DeviceCache.cs
@@ -11,6 +11,7 @@ namespace HidWizards.UCR.Core.Models
         public string DeviceHandle { get; set; }
         public int DeviceNumber { get; set; }
         public List<DeviceBindingNode> DeviceBindingMenu { get; set; }
+        public bool Blockable { get; set; }
 
     }
 }