+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Button to Filter`: Allows toggling of a filter using a button
 - `Axis to Filter`: Allows toggling of a filter using an axis
 - Added device cache allowing configuration of disconnected devices
+- Added per input blocking for supported providers
 - Provider Report now contains `ErrorMessage` property. If the provider is not live, this should contain a string indicating why
 - [Tobii Provider] If IsLive is false, now reports reason in ErrorMessage
 - [SpaceMouse provider] IsLive is always true, as HID is always present